added ROS state transmission for VRPuppet
