+++ Assets/Scripts/ROS/VRPuppetStateTransmissionServiceRequest.cs.meta
@@ -1,5 +1,5 @@
 fileFormatVersion: 2
-guid: 8fe71f4794c7e06448be92ed37b867f9
+guid: 40181ffb9cc474183a58856308637321
 MonoImporter:
   externalObjects: {}
   serializedVersion: 2
